refactor: restructure into compatibility file
